I found some errors using negative sign that were caused by std::strtold. To fix them tool now uses std::from_chars which cant handle Inf and nan, so they are handled explicitly. Now entries like -0 or -inf are handled correctly.
IEEE 745 floating point tool redesign modeled after 'float toy' web app (http://evanw.github.io/float-toy/)

Streamlined output using colors and compact layout which can be further simplified.
Chosen display mode (detailed or simplified) is automatically saved and set on new sessions.
Edit the binary bits, the integer hexadecimal or the floating point decimal values and the entire app will update with the change.
Supports the main IEEE745 standard formats (half, single and double precision) together with custom formats of size <= 64 bits.
Each format choice uses and displays the number of significant decimal digits defined by the mantissa size.
Added labels to identify the location of each bit box inside the binary representation.
Satisfies round trip / idempotent (reproducing) conversion property
Added theme colors, radio buttons for display mode and a clear button that resets the tool.
Removed previously and incorrectly added locale translations to various labels and languages
Attempted to adhere to code style formatting using existing code as example.
An effort was made to use preferred variable types and functions from std namespace when appropiate.
Attempted to document code using comments.

Not implemented / left to complete at a later time

Arbitrary width and precision formats.
Extended precision formats.
Shortest string property.
hexadecimal floating point display and conversions.

-----------------------------------
Old message from first pr commit
----------------------------------

Fixed a number of errors and added some features to the IEEE745 tool. There are no issues reporting these errors or feature request for the added features.

Errors:

The result written as a hexadecimal number had a few problems. The simpler way to note a problem is to set the least significand bit of the mantissa to 1 and leave the rest of the bits as zero. The resulting hexadecimal is shown as 2 instead of 1. If all the bits of the half precision case are set to 1 the resulting hexadecimal representation is expected to be 0xFFFF, but tool displayed 0x1FFFE. The total number of hexadecimal digits is determined by the chosen precision and should always be 4 for half precision (2 bytes), 8 for single precision (4 bytes) and 16 for double precision (8 bytes). Currently, the previous highest choice for precision determines the number of hex digits displayed, so choosing half precision after double did not change the resulting hexadecimal representation.

The formula for the exponent calculation was incorrect which resulted in entries like '2^(0-15) = 0'. This problem prevented subnormal numbers ( numbers with zero actual exponent and non-zero mantissa) from being displayed.

In addition to fixing the errors above the following new features are introduced.

The tool allows the user to specify the number of binary significand digits also known as the number's precision to values other than those specified in the standard.
When the value of the option that selects the desired numerical precision is changed, the display of all decimal numbers changes to reflect the new specified precision. The number of decimal significand digits that corresponds to the entered binary precision is obtained using methods described in the standard and/or related publications.

The tool now has, in addition to the decoder, the ability to encode IEE754 numbers and also non-standard formats.
Decimal value can be edited and all changes are propagated back to the rest of the interactive portions of the user interface of the tool. Valid entries include Infs and NaNs, denormals, and normals. The tool displays the group the current entry belongs to. Invalid entries are ignored. The enter key must be pressed to submit encoding requests.

A clear button that turns all the bits of the entry to zero was added for convenience especially for the larger precision choices. Appropriate localizations for the button message together with missing localizations for the tool description are included in this commit.