impr: clarify portal error message
`xdg-desktop-portal-wlr` is not relevant because it does not provide the
FileChooser interface.

`xdg-desktop-portal` needs the `WAYLAND_DISPLAY` env var on wayland and the
`DISPLAY` and `XAUTHORITY` env vars on xorg. If the dbus user session bus is
not run in a way that it already gets those variables, they must be given via
`dbus-update-activation-environment(1)`.

fixes #723