fix: Wrong query path when saving the layout
The layout manager incorrectly queried whether the empty path (effectively
the current working directory) is writable before saving the layout
(not "layouts" directories).

It resulted in (at least in Windows the author has tested) a failure to
save a layout when ImHex is installed as non-portable builds because:

1.  The current working directory is usually the one of
    the executable (imhex-gui.exe) and
2.  That directory (`C:\Program Files\ImHex` by default) is usually
    not writable unless ImHex is executed with an Administrator
    privilege.

This commit fixes this issue by querying "layouts" directories.