feat: Implement cross-platform DPI scaling
This replaces the Windows-only code with code that should work on
all the platforms. Tested on Linux/Wayland.

The “Scaling” option is now just treated like a normal UI zoom
control for users who want a larger UI. This means “Native” scaling
option is removed now since that does not mean anything.