Fix: fixing the mysterious 3d visualizer slows when axes are not drawn (issue #1663). It is not clear what is causing this bug. There is an optimization at the 2-d level that skips frames that are repeating the exact same draw calls that could be related but no reasonably detailed explanation as to how that is causing lag has been put forth.
The bug exhibits itself as a frame-locking of all windows according to their draw-lists updating frequency. If there are changes that occur 30 times a second anf this is the fastest occurring change then no windows will be allowed to update at a faster rate. The other major key is that only changes that are created using draw-list will be taken into account to set the locking rate. This explains why the bug is not present when the axes are shown since part of the axes are also the text labels "x","y" amd "z". Since the labels are constantly moving when rotating or translating the model then the draw-lists for these labels become the fastest changing ones. Note that the changes don't need to be visible, they just need to have a draw-list that updates at the desired rate.

To be clear about what this is, this is a fix for the lag seen under certain circumstances when using the 3d-visualizer. What this is not, is a fix for any or all underlying issues that may or may not exist in the code that affects rate locking or rate synchronization in ImHex. In that context it is not a workaround as it addresses the main issue (the lag) by making sure that there is a 2d draw-list associated with the display which is what is causing this bug (this i to say this fixes a bug of the 3d visualizer fixed using code from 3d_model.cpp , not a bug in windows.cpp which has code I am not familiar with). Adding a few vertices and indices to lists of potentially thousands of them is not going to have a measurable impact and honestly, the 3d -visualization tool is not sophisticated enough for us to worry about its performance after adding a letter or two.
The fix is very simple. We are already drawing 2d components in the 3d view when the axes are turned on, so adding three numbers for the x y and scroll numbers in the same color as the background is not a hack, it is the only way to fix the problem from within the visualizer code. That also means that it is not a workaround for the bug in question.

If there is a bug deep in the rendering code of ImHex then that bug needs to be found first before it can be fixed. Turning off the entire 2d optimization to avoid the lag would be considered a workaround for that bug. What needs to be done to fix the bug is to find a way to incorporate the 3d window into the 2d optimization. At any rate, that's a separate bug that may or may not be related to the one being fixed here. Additionally, the changes proposed here will not affect changes to fix the other bug if such a bug does in fact exist because the 3d visualizer code is at another separate level.

Without the intervention of those who know the rendering code well enough to make changes without creating bigger problems than the ones being solved, this is the best solution I can offer (and yes, it completely removes the lag)