fix: F3 failed to locate the correct set of matches if file was edited after doing a search.
The bug can be reproduced as follows:
1) Using ctrl-F do a search for a term that occurs several times in the file. and press f3 to visit them.
2) go back to the fop of the file and insert 3 blank lines.
3) pressing f3 once or many times will not find the term entered above.

The reason for this bug is that the positions of the matches are not being reset when changes can potentially move them. The fix consists on resetting the search locations when changing the contents of the file and redoing the search after the changes are made. The bug was specially problematic when doing replace because the replacement position would be identified as a match. This PR fixes replace as well.